requirements.txt
numpy>=1.24.0
torch>=2.0.0
matplotlib>=3.7.0
scipy>=1.10.0
flask>=2.3.0
flask-cors>=4.0.0
yfinance>=0.2.0
pandas>=2.0.0
pandas-ta>=0.3.14b0
nltk>=3.8.0
feedparser>=6.0.0
requests>=2.31.0
openai>=1.0.0
groq>=0.4.0
python-dotenv>=1.0.0

